Screeshot/requirements.txt
streamlit
pillow
easyocr
numpy
opencv-python-headless
torch
torchvision
sentence-transformers
scikit-learn